+++ game/board.py
@@ -13,10 +13,10 @@ class Board:
             for _ in range(8):
                 col.append(None)
             self.__positions__.append(col)
-        self.__positions__[0][0] = Rook("BLACK") # Black
-        self.__positions__[0][7] = Rook("BLACK") # Black
-        self.__positions__[7][7] = Rook("WHITE") # White
-        self.__positions__[7][0] = Rook("WHITE") # White
+        self.__positions__[0][0] = Rook("WHITE")
+        self.__positions__[0][7] = Rook("WHITE") 
+        self.__positions__[7][7] = Rook("BLACK") 
+        self.__positions__[7][0] = Rook("BLACK") 
         
         # Colocar los peones negros
         for i in range(8):
@@ -53,4 +53,20 @@ class Board:
             return None
         return piece.color
     
-board=Board()
+    
+    def show_board(self):
+        board_str = ""
+        for row in range(8):
+            for col in range(8):
+                piece = self.__positions__[row][col]
+                if piece is None:
+                    board_str += ". "
+                else:
+                    board_str += piece.__str__() + " "
+            board_str = board_str.strip() + "\n"
+        return board_str.strip()
+
+
+if __name__ == "__main__":
+    board = Board()
+    print(board.show_board())