+++ game/rook.py
@@ -7,7 +7,7 @@ class Rook(Piece):
         else:
             return "♖"
 
-
+  
     # def valid_positions(
     #     self,
     #     from_row,