+++ test/test_board.py
@@ -14,10 +14,8 @@ class TestBoard(unittest.TestCase):
         
     def test_initial_positions_rook(self):
         # Verificar que las torres negras estén en las posiciones correctas
+        self.assertEqual(self.board.__positions__[0][0].color, "WHITE")
 
-        self.assertEqual(self.board.__positions__[0][0].color, "BLACK")
-
-        
     def test_empty_positions(self):
         # Verificar que las posiciones vacías sean None
         for row in range(2, 6):
@@ -25,10 +23,20 @@ class TestBoard(unittest.TestCase):
                 self.assertIsNone(self.board.get_piece(row, col))
                 
     def test_positions(self):
-        self.assertEqual(self.board.get_piece(0, 0), "BLACK")
-        
-                
-        
+        self.assertEqual(self.board.get_piece(0, 0), "WHITE")
+       
+       
+    def test_show_board(self):
+        expected_output = "♜ ♞ ♝ ♛ ♚ ♝ ♞ ♜\n" \
+                          "♟ ♟ ♟ ♟ ♟ ♟ ♟ ♟\n" \
+                          ". . . . . . . .\n" \
+                          ". . . . . . . .\n" \
+                          ". . . . . . . .\n" \
+                          ". . . . . . . .\n" \
+                          "♙ ♙ ♙ ♙ ♙ ♙ ♙ ♙\n" \
+                          "♖ ♘ ♗ ♕ ♔ ♗ ♘ ♖"
+        self.assertEqual(expected_output, self.board.show_board())
+               
 
 if __name__ == '__main__':
     unittest.main()